Type changed from integer to float (OrderInventory.releaseQty)

+++ server/graphql/types/order-inventory/order-inventory-patch.ts
@@ -15,7 +15,7 @@ export const OrderInventoryPatch = gql`
     productName: String
     batchId: String
     packingType: String
-    releaseQty: Int
+    releaseQty: Float
     releaseWeight: Float
     remark: String
     issue: String
